# api/openapi.yaml
openapi: 3.1.0

info:
  version: 1.0.0
  title: Vixar API
  description: Public API for Vixar

servers:
  - url: http://localhost:8000/api/v1/

paths:
  /api/v1/token/search:
    post:
      tags: [token]
      summary: Get info for specified token
      operationId: searchTokenInfo
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchTokenInfoRequest'
      responses:
        '200':
          description: Successfully retrieved interest data
          content:
            application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/TokenInfo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Token not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/vk/callback:
    post:
      tags: [auth]
      summary: VK oauth callback, used to get tokens and user info from vk
      operationId: vkAuthCallback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VkAuthCallbackRequest'
      responses:
        '200':
          description: Successfully authenticated with vk
          headers:
            Set-Cookie:
              description: Session cookie
              schema:
                type: string
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/VkAuthCallbackResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '406':
          description: Not acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/vk/register:
    post:
      tags: [auth]
      summary: VK oauth register, used to register new users via vk
      operationId: vkAuthRegister
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VkAuthRegisterRequest'
      responses:
        '200':
          description: Successfully registered with vk
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (should be session for vk user registration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict - user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Logout user
      operationId: logoutUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully logged out
          headers:
            Set-Cookie:
              description: Clear session cookie
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id

  schemas:
    SearchTokenInfoRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 4
          maxLength: 255
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
      required: [token, start]
    TokenInfo:
      type: object
      properties:
        token:
          type: string
        records:
          type: array
          items:
            $ref: '#/components/schemas/TokenRecord'
      required: [token, records]
    TokenRecord:
      type: object
      properties:
        timestamp:
          type: string
        features:
          type: object
          properties:
            interest:
              type: integer
              format: int64
            interest_normalized:
              type: number
              format: float64
            sentiment:
              type: integer
              format: int16
          required: [interest, interest_normalized, sentiment]
      required: [timestamp, features]
    VkAuthCallbackRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 256
        state:
          type: string
          minLength: 1
          maxLength: 256
        code_verifier:
          type: string
          minLength: 1
          maxLength: 256
        device_id:
          type: string
          minLength: 1
          maxLength: 256
        redirect_uri:
          type: string
          minLength: 1
          maxLength: 256
      required: [code, state, code_verifier, device_id, redirect_uri]
    VkAuthCallbackResponse:
      type: object
      properties:
        user_exists:
          type: boolean
        username:
          type: string
        email:
          type: string
        vkid:
          type: integer
          format: int64
      required: [user_exists, username, email, vkid]
    VkAuthRegisterRequest:
      type: object
      properties:
        email:
          type: string
          minLength: 4
          maxLength: 64
        username:
          type: string
          minLength: 4
          maxLength: 64
        vkid:
          type: integer
          format: int64
      required: [email, username, vkid]

    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]